object/requirements.txt
python-dotenv
numpy
discord
requests
pillow
download
dlib
uuid
opencv-contrib-python
